rename and number figures
