+++ nuc_morph_analysis/analyses/evaluate_filters_and_outliers/figure_s7_workflow.py
