+++ nuc_morph_analysis/analyses/volume/figure_4_s8_workflow.py
@@ -21,12 +21,12 @@ df = global_dataset_filtering.load_dataset_with_features()
 df_full = filter_data.all_timepoints_full_tracks(df)
 df_track_level_features = filter_data.track_level_features(df_full)
 
+# %% MAIN FIGURE 4
+
 # %%
 # Plot all tracks all datasets synchronized in real or normalized time
 for time in ["sync_time_Ff", "normalized_time"]:
     plot_tracks.plot_all_tracks_synchronized(df_full, figdir, "volume", time)
-
-# %% MAIN FIGURE PANELS
 
 # %%
 # Plot single sample track from same nucleus as in formation/breakdown figure
@@ -73,7 +73,7 @@ plt.xlabel(f"{label} {units}")
 plt.ylabel("Counts")
 save_and_show_plot(f"{figdir}/hist_volume_fold_change_BC")
 
-# %% SUPPLEMENTAL FIGURE PANELS
+# %% SUPPLEMENTAL FIGURE S8
 
 # %%
 # Plot mean of all volume and fold-change tracks with 5th to 95th percentile shaded
